remove graph
