+++ contracts/index_executor/src/steps/swap.rs
@@ -113,7 +113,7 @@ impl Runner for SwapStep {
 
         // Query off-chain indexer directly get the execution result
         let chain = &context
-            .graph
+            .registry
             .get_chain(self.chain.clone())
             .ok_or("MissingChain")?;
         let account = match chain.chain_type {
